+++ .eslintrc.cjs
@@ -78,12 +78,7 @@ module.exports = {
         "import/order": [
           "warn",
           {
-            groups: [
-              "builtin",
-              "external",
-              "internal",
-              ["parent", "sibling", "index"],
-            ],
+            groups: ["builtin", "external", "internal", ["parent", "sibling", "index"]],
             "newlines-between": "always",
             alphabetize: { order: "asc", caseInsensitive: true },
           },
@@ -106,6 +101,6 @@ module.exports = {
     },
   ],
   globals: {
-    shopify: "readonly"
+    shopify: "readonly",
   },
 };